Rename Number 3.ipynb to code 1.ipynb
We have selected a commodity (say coffee) for a given market (say Kalpetta of state Kerala) and computed the moving average of the modal prices for a 7-day window. Then the difference of the Modal Price (at that day) and Moving Average (from 'day-3' to 'day+3' ie 7 data points ) is computed. The Coefficient of Variation (CV) of Prices of that commodity (Coffee) in that market (Kalpetta) is computed. 
If the absolute value of the difference exceeds CV, then the data point is classified as an outlier and flag has been marked as 1. Similarly for all the commodities of all markets in the given year.